+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fixed: Bug with undefined content variable with IN operation (issue #501)
 * Fixed: Bug with wrong realizaion of REPLACE INTO (issue #505)
 * Added: >>,<<,&,| - binary operations
+* Added: || - string concatenation
 
 ### 0.2.1 "Rodos" (28.09.2015)
 * Added: AlaSQL CLI: Support for --version flag